Cory Davis  <cory@met.ed.ac.uk>
	* arts-1-1-734

	* src/m_continua_ckd.cc: arts not compiling on gcc 2.96, so I added
	#include <cmath>, which seems to have worked.

	* src/m_abs_o2_models.cc: changing "AN_r = abs(N_r)-1" to
	"AN_r = abs(Numeric(N_r))-1" and others like it fixed compilation
	errors (ambiguity with overloaded function abs) with gcc 2.96.  Feel
	free to flame me if these were not the right fixes.


git-svn-id: https://www.sat.ltu.se/svn/rt/arts/trunk@3814 aaf1aab0-4228-0410-ad68-8dceda47f409

+++ src/m_abs_o2_models.cc
@@ -1162,11 +1162,11 @@ void Zeeman_o2_line_splitting(// Output:
     { 
       // Assigning a value of the total angular momentum N of the lower  
       // level for a N+ Zeeman transition (N -> N+1).
-      AN_r = abs(N_r);
+      AN_r = abs(Numeric(N_r));
       
       // Assigning a value of the number of the split components for each 
       // of the three possible polarizations of a N+ Zeeman transition.
-      BN_r = 2*abs(N_r)+1;
+      BN_r = 2*abs(Numeric(N_r))+1;
       
       //! Resize of the frequency shift matrix 
       // for a N+ Zeeman transition.
@@ -1180,11 +1180,11 @@ void Zeeman_o2_line_splitting(// Output:
     {
       // Assigning a value of the total angular momentum N of the lower 
       // level for a N- Zeeman transition (N -> N-1).
-      AN_r = abs(N_r)-1;
+      AN_r = abs(Numeric(N_r))-1;
       
       // Assigning a value of the number of the split components for each
       // of the three possible polarizations of a N- Zeeman transition.
-      BN_r = 2*abs(N_r)-1;
+      BN_r = 2*abs(Numeric(N_r))-1;
       
       
       // Resize of the frequency shift matrix
